Move files to recommended location in FE source code
